rtrtrt

+++ src/common.h
@@ -68,7 +68,7 @@ typedef struct {
 
     uint16_t adc; //последние значение с adc
 
-    uint16_t power; //текущая мощность
+    uint8_t power; //текущая мощность
     uint16_t temp; //текущая температура тены
     uint16_t temp_need; //требуемая температура
